Added button to create a new collecting session
It works OK, but I have to reload the page manually, and that destroys the state
that would make the dialog more usable.

The JSON round-trip has reformatted a few parts of `lepid.json`, but that should
be a one-time thing.

I added several more empty sessions for June 1977 during testing.