Separated modes
VIEWING and COLLECTING modes are working.  V supports "id" as a
query parameter, and opens to that specimen correctly.
It still needs a more general search feature, and perhaps some
specific ones (e.g. "genus=...").

Storage lifecycle is still not worked out, but I can see the collection
in all three modes, and I was able to save data to local and
load from local, as well as discarding changes by loading from
the resource again.

I have switched to an ID rather than an object reference as the selection
state.  The selected specimen and its container are derived signals.

I also did a big refactor to break the app up into several files, so diffs
are going to be hard, unfortunately.

Everything seems to be working.